Move ObservationsDownloadForm to Views/observations/downloads/form
Single use cluster boundary case: previously rendered from both
`Observations::DownloadsController` (its own controller) and
`SpeciesLists::DownloadsController` (downloading the obs in a list).
Move to its natural home under the observations cluster and have
species_lists/downloads reference it cross-namespace.

  Components::ObservationsDownloadForm
    → Views::Controllers::Observations::Downloads::Form

Form id changes per the all-segments heuristic:
  observations_download_form → observation_download_form
(only one downstream test referenced it.)

`species_lists/downloads/new.html.erb` keeps its render call
intact, just qualifies the class to the new namespace:
  render(Views::Controllers::Observations::Downloads::Form.new(...))

45 view tests + 282 controller tests pass. Lint clean.
