+++ app/views/controllers/species_lists/downloads/new.html.erb
@@ -11,7 +11,7 @@ query_param = q_param(@query)
 <%= render(partial: "species_lists/downloads/form_species_list_report",
            locals: { query_param: query_param }) %>
 
-<%= render(Components::ObservationsDownloadForm.new(
+<%= render(Views::Controllers::Observations::Downloads::Form.new(
       query_param: query_param,
       format: @format,
       encoding: @encoding
